[DF-6560] Compiled AssetCacheServer for win32
Former-commit-id: d22ce06cb3029ffe780e1cf4f3ffae8467aee769

+++ Sources/Tools/AssetCache/CacheDB.cpp
@@ -126,7 +126,7 @@ void CacheDB::Load()
     
     usedSize = header->GetUInt64("usedSize");
     auto cacheSize = header->GetUInt64("itemsCount");
-    fullCache.reserve(cacheSize);
+    fullCache.reserve(static_cast<size_t>(cacheSize));
 
     ScopedPtr<KeyedArchive> cache(new KeyedArchive());
     cache->Load(file);
